requirements.txt
django==5.0.3
psycopg2-binary==2.9.9
djangorestframework==3.15.1
pillow==10.2.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
celery==5.3.6
django-celery-beat==2.6.0
redis==5.0.3
coverage==7.4.4
drf-yasg==1.21.7
django-cors-headers==4.3.1
python-dotenv==1.0.1